rename class PubNubNS to PubNub
